Restructuring for clarity and add ListCoursesServlet
