+++ project/src/main/java/com/google/autograder/servlets/helpers/API.java
@@ -1,0 +1,5 @@
+package com.google.autograder.servlets.helpers;
+
+public final class API {
+    
+}
